make entity source hand;ler more generic
